requirements.txt
torch>=2.1.0
transformers>=4.51.3
accelerate>=0.25.0
diffusers @ git+https://github.com/huggingface/diffusers.git
Pillow>=10.0.0
runpod>=1.6.0
requests>=2.31.0
numpy>=1.24.0
safetensors>=0.4.0
sentencepiece>=0.1.99
protobuf>=4.25.0
